requirements.txt
pymupdf
python-docx
sentence-transformers
faiss-cpu
numpy
requests
fastapi
uvicorn
pydantic